chore(alert-banner): refactor AlertBanner to extend AlertBannerBase and add variant handling
