+++ packages/alert-banner/package.json
@@ -25,6 +25,10 @@
             "default": "./src/index.js"
         },
         "./package.json": "./package.json",
+        "./src/AlertBannerBase.js": {
+            "development": "./src/AlertBannerBase.dev.js",
+            "default": "./src/AlertBannerBase.js"
+        },
         "./src/AlertBanner.js": {
             "development": "./src/AlertBanner.dev.js",
             "default": "./src/AlertBanner.js"
