+++ packages/alert-banner/src/AlertBanner.ts
@@ -12,18 +12,13 @@
 import {
     CSSResultArray,
     html,
-    PropertyValues,
-    SpectrumElement,
     TemplateResult,
 } from '@spectrum-web-components/base';
-import { property } from '@spectrum-web-components/base/src/decorators.js';
 import '@spectrum-web-components/button/sp-close-button.js';
 import '@spectrum-web-components/icons-workflow/icons/sp-icon-alert.js';
 import '@spectrum-web-components/icons-workflow/icons/sp-icon-info.js';
 import styles from './alert-banner.css.js';
-
-const VALID_VARIANTS = ['neutral', 'info', 'negative'];
-export type AlertBannerVariants = (typeof VALID_VARIANTS)[number];
+import { AlertBannerBase } from './AlertBannerBase.js';
 
 /**
  * @element sp-alert-banner
@@ -33,72 +28,12 @@ export type AlertBannerVariants = (typeof VALID_VARIANTS)[number];
  *
  * @fires close - Announces the alert banner has been closed
  */
-export class AlertBanner extends SpectrumElement {
+export class AlertBanner extends AlertBannerBase {
     public static override get styles(): CSSResultArray {
         return [styles];
     }
 
-    /**
-     * Controls the display of the alert banner
-     *
-     * @param {Boolean} open
-     */
-    @property({ type: Boolean, reflect: true })
-    public open = false;
-
-    /**
-     * Whether to include an icon-only close button to dismiss the alert banner
-     *
-     * @param {Boolean} dismissible
-     */
-    @property({ type: Boolean, reflect: true })
-    public dismissible = false;
-
-    /**
-     * The variant applies specific styling when set to `negative` or `info`;
-     * `variant` attribute is removed when it's passed an invalid variant.
-     *
-     * @param {String} variant
-     */
-    @property({ type: String })
-    public set variant(variant: AlertBannerVariants) {
-        if (variant === this.variant) {
-            return;
-        }
-        const oldValue = this.variant;
-
-        if (this.isValidVariant(variant)) {
-            this.setAttribute('variant', variant);
-            this._variant = variant;
-        } else {
-            this.removeAttribute('variant');
-            this._variant = '';
-
-            if (window.__swc.DEBUG) {
-                window.__swc.warn(
-                    this,
-                    `<${this.localName}> element expects the "variant" attribute to be one of the following:`,
-                    'https://opensource.adobe.com/spectrum-web-components/components/alert-banner/#variants',
-                    {
-                        issues: [...VALID_VARIANTS],
-                    }
-                );
-            }
-        }
-        this.requestUpdate('variant', oldValue);
-    }
-
-    public get variant(): AlertBannerVariants {
-        return this._variant;
-    }
-
-    private _variant: AlertBannerVariants = '';
-
-    protected isValidVariant(variant: string): boolean {
-        return VALID_VARIANTS.includes(variant);
-    }
-
-    protected renderIcon(variant: string): TemplateResult {
+    protected override renderIcon(variant: string): TemplateResult {
         switch (variant) {
             case 'info':
                 return html`
@@ -113,29 +48,6 @@ export class AlertBanner extends SpectrumElement {
                 `;
             default:
                 return html``;
-        }
-    }
-
-    private shouldClose(): void {
-        const applyDefault = this.dispatchEvent(
-            new CustomEvent('close', {
-                composed: true,
-                bubbles: true,
-                cancelable: true,
-            })
-        );
-        if (applyDefault) {
-            this.close();
-        }
-    }
-
-    public close(): void {
-        this.open = false;
-    }
-
-    private handleKeydown(event: KeyboardEvent): void {
-        if (event.code === 'Escape' && this.dismissible) {
-            this.shouldClose();
         }
     }
 
@@ -160,17 +72,5 @@ export class AlertBanner extends SpectrumElement {
                     : html``}
             </div>
         `;
-    }
-
-    protected override updated(changes: PropertyValues): void {
-        super.updated(changes);
-
-        if (changes.has('open')) {
-            if (this.open) {
-                this.addEventListener('keydown', this.handleKeydown);
-            } else {
-                this.removeEventListener('keydown', this.handleKeydown);
-            }
-        }
     }
 }
